requirements.txt
imutils
numpy
opencv-python
scipy
lz4
tqdm
customtkinter
uiautomator2
pponnxcr
Pillow
psutil
pywin32